Cambios en funciones de preferencia BD, cambios en proyecto Pochoclo quitando las creaciones de tablas desde el mismo. Tambien en preferencias para listar la pref. de un Usuario.

+++ PochocloCritics/src/main/java/org/pochocloapps/pochoclocritics/controladores/DirectoresControlador.java
@@ -34,7 +34,7 @@ public class DirectoresControlador {
     }
     
     public void borrar(Context ctx) throws SQLException, DirectorNoEncontradoException{
-        directoresRepositorio.borrar(directoresRepositorio.obtener(ctx.pathParam("idDirector", Long.class).get()));
+        directoresRepositorio.borrar(directoresRepositorio.obtener(ctx.pathParam("idDirector", Integer.class).get()));
         ctx.status(204);
         
     }
